refactor: blockchain file
