+++ src/core/types/index.ts
@@ -1,7 +1,7 @@
 export * from './Alerts';
+export * from './Blockchain';
 export * from './Config';
 export * from './Container';
-export * from './Ethereum';
 export * from './IronBank';
 export * from './Lab';
 export * from './Modals';
